this fucking script ughh

+++ histogramEqualization.py
@@ -21,7 +21,7 @@ for i in range (0,y):
     fda.append(0)
     for j in range (0,i+1):
         fda[i] += p[j]
-    fda[i]= np.around(fda[i]*(y-1))
+    fda[i]=int(np.around(fda[i]*(y-1)))
 
 print
 print
